# Shader processor optimization