@Popov72 2023-05-29
> I see that ShaderProcessor does not have own test file. Maybe it possible to test with screenshot tests, but how it runs now?

It's automatically tested by all the visualization tests because they all compile shaders.